Shuffle some logic around
